* upstream/trunk: (51 commits)
  Register rewrite Update, bug fixes and allow mouse to change reconcile flag.
  Register rewrite Update, bug fixes and allow mouse to change reconcile flag.
  Cleanup non-relevant comments introduced with r22825
  Support for querying Bitcoin/Litecoin prices via F::Q's MtGox module Author:    Sam Morris <sam@robots.org.uk> BP
  Fix a typo in both source and po files.
  Update Italian translation.
  Add script to generate a list of changes in html format between two revisions based on git log
  Add code to generate ChangeLog from git
  Win32 git: fetch potentially new tags before running the tags build
  Win32 fix path to git build script for tags
  Win32: build 2.5 tagged releases from git instead of svn
  Fix charts on Windows
  Bug #700197 - Critical SQL backend errors when creating customers via python bindings
  Round properly when computing fraction of the transaction's value due to a given split. Also use GNC-DENOM-AUTO instead of a constant zero when appropriate.
  Bug #682800 Generated balances on Report different than calculated balances on Ledger when using "open subaccounts"
  Bug #622778 Miscalculation in cashflow reports - Step 03
  Bug #622778 Miscalculation in cashflow reports - Step 02
  Bug #622778 Miscalculation in cashflow reports - Step 01
  Bug #584869 net change line in general journal report broken
  Bug #589865 In the report "Budget flow" period it doesn't work ok
  ...

+++ packaging/win32/build_tags_git.sh
@@ -39,6 +39,7 @@ $GIT_CMD pull
 ################################################################
 # determine if there are any new tags since the last time we ran
 #
+$GIT_CMD fetch -t
 
 # If we don't have a tagfile then start from 'now'
 tagfile=tags_git
@@ -68,6 +69,17 @@ mv -f ${tagfile}.new ${tagfile}
 for tag_rev in $tags ; do
   tag=${tag_rev#*/}
   tag=${tag%/*}
+  
+  # Git builds are only supported from 2.5 up
+  tag_major=${tag%%.*}
+  tag_tmp=${tag#*.}
+  tag_minor=${tag_tmp%%.*}
+  major_minor=$(( $tag_major*100 + $tag_minor ))
+  if (( $major_minor < 205 ))
+  then
+     continue
+  fi
+  
   tagbasedir=/c/soft/gnucash-${tag}
   tagdir=${tagbasedir}/gnucash
   rm -fr $tagbasedir
